+++ src/Audit.EntityFramework/ConfigurationApi/ContextSettingsConfigurator.cs
@@ -33,12 +33,7 @@ namespace Audit.EntityFramework.ConfigurationApi
             Configuration.SetExcludeTransactionId<T>(exclude);
             return this;
         }
-        
-        public IContextSettingsConfigurator<T> EarlySavingAudit(bool earlySaving = true)
-        {
-            Configuration.SetEarlySavingAudit<T>(earlySaving);
-            return this;
-        }
+
 #if EF_FULL
         public IContextSettingsConfigurator<T> IncludeIndependantAssociations(bool include = true)
         {