+++ src/Audit.EntityFramework/ConfigurationApi/EfSettings.cs
@@ -24,7 +24,6 @@ namespace Audit.EntityFramework.ConfigurationApi
         public Func<Type, bool> IncludedTypesFilter { get; set; }
         public Dictionary<Type, EfEntitySettings> EntitySettings { get; set; }
         public bool? ExcludeTransactionId { get; set; }
-        public bool? EarlySavingAudit { get; set; }
 #if EF_FULL
         public bool? IncludeIndependantAssociations { get; set; }
 #endif