+++ src/Audit.EntityFramework/DefaultAuditContext.cs
@@ -29,7 +29,6 @@ namespace Audit.EntityFramework
         public virtual void OnScopeSaving(IAuditScope auditScope) { }
         public virtual void OnScopeSaved(IAuditScope auditScope)  { }
         public bool ExcludeTransactionId { get; set; }
-        public bool EarlySavingAudit { get; set; }
 #if EF_FULL
         public virtual bool IncludeIndependantAssociations { get; set; }
 #endif